+++ jest.github.config.ts
@@ -4,6 +4,7 @@ import baseConfig from './jest.config';
 const config: JestConfigWithTsJest = Object.assign(baseConfig, {
   collectCoverage: true,
   coverageReporters: ['lcov'],
+  logHeapUsage: true,
   reporters: ['default', 'github-actions']
 } as JestConfigWithTsJest);