+++ src/features/emojis/emojis.spec.ts
@@ -14,18 +14,6 @@ jest.mock('@ta-x-utilities', () => {
   };
 });
 
-jest.mock('emoji.json', () => {
-  /* eslint-disable @typescript-eslint/no-var-requires */
-  const compress = require('compress-json').compress;
-  const groupEmojis = require('@ta-x-build-helpers').groupEmojis;
-  const emojiJson = fs.readFileSync(require.resolve('emoji.json'), 'utf8');
-  /* eslint-enable @typescript-eslint/no-var-requires */
-
-  const mappedEmojis = groupEmojis(emojiJson);
-  const compressedEmojis = compress(mappedEmojis);
-  return compressedEmojis;
-});
-
 describe('emojis', () => {
   beforeEach(() => {
     setHtml('@ta-x-jest-views/empty.html');