+++ src/types/models/gamertag-nation-opts.ts
@@ -1,0 +1,3 @@
+import { ImportableGuideOptions } from './importable-guide-opts';
+
+export type GamertagNationOptions = ImportableGuideOptions;