+++ src/types/models/index.ts
@@ -1,2 +1,3 @@
 export * from './importable-guide-opts';
 export * from './resero-network-opts';
+export * from './gamertag-nation-opts';